+++ src/components/avatar/index.js
@@ -1,5 +1,6 @@
 import Avatar from './Avatar';
+import AvatarInitials from './AvatarInitials';
 import AvatarStack from './AvatarStack';
 
 export default Avatar;
-export { Avatar, AvatarStack };
+export { Avatar, AvatarInitials, AvatarStack };